requirements.txt
pynsq===0.9.1